🚚 api: move csv utils to results context
