+++ api/src/certification/results/infrastructure/utils/csv/certification-results/CertificationResultsCsvValues.js
