+++ api/src/certification/results/infrastructure/utils/csv/certification-results/SessionCertificationResultsCsvBuilder.js
