+++ api/src/certification/results/infrastructure/utils/csv/certification-results/get-division-certification-results-csv.js
@@ -1,6 +1,6 @@
 import dayjs from 'dayjs';
 
-import { getCsvContent } from '../../../../../src/shared/infrastructure/utils/csv/write-csv-utils.js';
+import { getCsvContent } from '../../../../../../shared/infrastructure/utils/csv/write-csv-utils.js';
 import { DivisionCertificationResultsCsvBuilder } from './DivisionCertificationResultsCsvBuilder.js';
 
 const getDivisionCertificationResultsCsv = async function ({ division, certificationResults, i18n }) {
