+++ api/tests/certification/results/integration/infrastructure/utils/csv/certification-results/get-division-certification-results-csv_test.js
@@ -1,13 +1,13 @@
 import dayjs from 'dayjs';
 
-import { getDivisionCertificationResultsCsv } from '../../../../../../lib/infrastructure/utils/csv/certification-results/get-division-certification-results-csv.js';
-import { AutoJuryCommentKeys } from '../../../../../../src/certification/shared/domain/models/JuryComment.js';
-import { domainBuilder, expect } from '../../../../../test-helper.js';
-import { getI18n } from '../../../../../tooling/i18n/i18n.js';
+import { getDivisionCertificationResultsCsv } from '../../../../../../../../src/certification/results/infrastructure/utils/csv/certification-results/get-division-certification-results-csv.js';
+import { AutoJuryCommentKeys } from '../../../../../../../../src/certification/shared/domain/models/JuryComment.js';
+import { domainBuilder, expect } from '../../../../../../../test-helper.js';
+import { getI18n } from '../../../../../../../tooling/i18n/i18n.js';
 const i18n = getI18n();
 const translate = i18n.__;
 
-describe('Integration | Application | UseCases | certification-results | get-division-certification-results-csv', function () {
+describe('Certification | Results | Integration | Infrastructure | Utils | certification-results | get-division-certification-results-csv', function () {
   context('#getDivisionCertificationResultsCsv', function () {
     context('when at least one candidate has passed a certification', function () {
       it('returns a csv without session information', async function () {
