+++ api/tests/certification/results/unit/infrastructure/utils/csv/certification-results/CertificationResultsCsvValues_test.js
@@ -1,8 +1,8 @@
-import { CertificationResult } from '../../../../../../lib/domain/models/CertificationResult.js';
-import { CertificationResultsCsvValues } from '../../../../../../lib/infrastructure/utils/csv/certification-results/CertificationResultsCsvValues.js';
-import { AutoJuryCommentKeys } from '../../../../../../src/certification/shared/domain/models/JuryComment.js';
-import { domainBuilder, expect } from '../../../../../test-helper.js';
-import { getI18n } from '../../../../../tooling/i18n/i18n.js';
+import { CertificationResult } from '../../../../../../../../lib/domain/models/index.js';
+import { CertificationResultsCsvValues } from '../../../../../../../../src/certification/results/infrastructure/utils/csv/certification-results/CertificationResultsCsvValues.js';
+import { AutoJuryCommentKeys } from '../../../../../../../../src/certification/shared/domain/models/JuryComment.js';
+import { domainBuilder, expect } from '../../../../../../../test-helper.js';
+import { getI18n } from '../../../../../../../tooling/i18n/i18n.js';
 
 describe('Unit | Infrastructure | Utils | Csv | CertificationResultsCsvValues', function () {
   let i18n, translate;
